+++ src/main/java/com/pokemonspring/service/PokemonAPIService.java
@@ -3,9 +3,9 @@ package com.pokemonspring.service;
 import com.pokemonspring.model.Pokemon;
 
 public interface PokemonAPIService {
-    Pokemon getPokemon(Long id);
+    Pokemon getPokemon(int id);
 
     Pokemon json2Pokemon(String json);
 
-    String getPokemonFromAPI(Long id);
+    String getPokemonFromAPI(int id);
 }
